+++ MsgPack5.H5/Internal/ArrayDataDecoder.cs
@@ -4,17 +4,17 @@ namespace MsgPack5.H5
 {
     internal sealed class ArrayDataDecoder
     {
-        private readonly Func<ulong, Type> _expectedTypeForIndex;
-        private readonly Action<ulong, object> _setterForIndex;
+        private readonly Func<uint, Type> _expectedTypeForIndex;
+        private readonly Action<uint, object> _setterForIndex;
         private readonly Func<object> _finalResultGenerator;
-        public ArrayDataDecoder(Func<ulong, Type> expectedTypeForIndex, Action<ulong, object> setterForIndex, Func<object> finalResultGenerator)
+        public ArrayDataDecoder(Func<uint, Type> expectedTypeForIndex, Action<uint, object> setterForIndex, Func<object> finalResultGenerator)
         {
             _expectedTypeForIndex = expectedTypeForIndex ?? throw new ArgumentNullException(nameof(expectedTypeForIndex));
             _setterForIndex = setterForIndex ?? throw new ArgumentNullException(nameof(setterForIndex));
             _finalResultGenerator = finalResultGenerator ?? throw new ArgumentNullException(nameof(finalResultGenerator));
         }
-        public Type GetExpectedTypeForIndex(ulong index) => _expectedTypeForIndex(index);
-        public void SetValueAtIndex(ulong index, object value) => _setterForIndex(index, value);
+        public Type GetExpectedTypeForIndex(uint index) => _expectedTypeForIndex(index);
+        public void SetValueAtIndex(uint index, object value) => _setterForIndex(index, value);
         public object GetFinalResult() => _finalResultGenerator();
     }
 }
